Update ReadingShoeDataTrainingModels_Xiaowei_v3.m
Best model so far after trial and error. A couple of things that helped based on 40 different combinations include (1) the addition of a bilstm layer as the first layer in the network, (2) the normalisation of the time feature, (3) the addition of the masks (this is even better than removing corrupt data from the training).